Struct rust_battleship::battleship_game::player::Player

pub struct Player {
    board_matrix: [[WaterSquare; 10]; 9],
    ships_vec: Vec<Ship>,
    player_signature: Players,
}

This is the struct that contains the data for each indiviual player

Fields
board_matrix: [[WaterSquare; 10]; 9]

A matrix of size BOARD_WIDTH x BOARD_HEIGHT to store which squares were hit and which ones weren’t

ships_vec: Vec<Ship>

A vector of the ship objects the player has on their board

player_signature: Players

The signature of the player, mostly used to let the front end know which player achieved a victory

Implementations

impl Player

pub fn new_player(ship_count: usize, player_sig: Players) -> Player

Static function that creates a new player object with the given parameters:

ship_count: the amount of ships that this player should have (currently unused)

player_sig: the signature of the player in the game, so the code can differentiate which player is which

pub fn place_ship(&mut self, starting_point: (u8, u8), is_vertical: bool)

Function to place a ship on this player, the size of the ship is automatically determined by how many ships are already placed on the board

starting_point: a tuple contiaining the point where the ship should begin

is_vertical: determines whether or not the ship should be vertical, if this is false, the ship will go to the right from starting_point, if it’s true, then it will go down

pub fn fire(&mut self, pos: (u8, u8)) -> FireState

Function for being hit on a board, this will be called when it’s the other player’s turn. This returns a FireState so the game knows if it needs to check for sinked ships

pos: a tuple containing the position on the board that we’re firing on

pub fn ship_index_at(&self, pos: (u8, u8)) -> Option<usize>

Function for returning if there’s a ship at a particular location, if there’s a ship, this returns Some(i) with i being the index of the ship at that location in the vector, if there’s no ship, None is returned instead

pos: the position to check for a ship at

pub fn all_ships_sunk(&self) -> bool

This checks if all the ships that this player has on their board have sunk, returns respective bool value

pub fn get_ship_count(&self) -> usize

Returns the amount of ships the player has on their board

pub fn get_sig(&self) -> &Players

Gets the player signature, mostly to know at the end which player has won the game

impl Player

pub fn print_ships(&self)

Prints the vectors of the ships the player has into the command line, mostly for debugging

pub fn print_board(&self)

Prints the board to show which squares are hit and which ones are not to the command line, mostly for debugging